% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Accuracy.R
\name{ATA.Accuracy}
\alias{ATA.Accuracy}
\title{Accuracy Measures for a ATA forecast model}
\usage{
ATA.Accuracy(object, out.sample = NULL)
}
\arguments{
\item{object}{An object of class \code{ATA} is required.}

\item{out.sample}{A numeric vector or time series of class \code{ts} or \code{msts} for out-sample.}
}
\value{
Matrix giving forecast accuracy measures.
}
\description{
Returns ATA(p,q,phi) applied to ATA \code{object}.
Accuracy measures for a forecast model
Returns range of summary measures of the forecast accuracy. If \code{out.sample} is
provided, the function measures test set forecast accuracy.
If \code{out.sample} is not provided, the function only produces
training set accuracy measures.
The measures calculated are:
\itemize{
		\item{MAE}	 : mean absolute error.
		\item{MSE}	 : mean square error.
		\item{RMSE}	 : root mean squared error.
		\item{MPE}	 : mean percentage error.
		\item{MAPE}	 : mean absolute percentage error.
		\item{sMAPE} : symmetric mean absolute percentage error.
		\item{MASE}	 : mean absolute scaled error.
		\item{OWA}	 : overall weighted average of MASE and sMAPE.
		\item{MdAE}	 : median absolute error.
		\item{MdSE}	 : median square error.
		\item{RMdSE} : root median squared error.
		\item{MdPE}	 : median percentage error.
		\item{MdAPE} : median absolute percentage error.
		\item{sMdAPE}: symmetric median absolute percentage error.
}
}
\examples{
ata.fit1 <- ATA(M3[[1899]]$x, h=18)
ATA.Accuracy(ata.fit1)
ATA.Accuracy(ata.fit1, M3[[1899]]$xx)
}
\references{
Hyndman, R.J. and Koehler, A.B. (2006) "Another look at measures
of forecast accuracy". \emph{International Journal of Forecasting},
\bold{22}(4), 679-688. Hyndman, R.J. and Athanasopoulos, G. (2014)
"Forecasting: principles and practice", OTexts. Section 2.5 "Evaluating forecast accuracy". \url{http://www.otexts.org/fpp/2/5}.
}
\seealso{
\code{forecast}, \code{stlplus}, \code{stR}, \code{\link[stats]{stl}}, \code{\link[stats]{decompose}}, \code{tbats}, \code{seasadj}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
\keyword{accuracy}
\keyword{ata}
\keyword{forecast}
\keyword{msts}
\keyword{ts}
