% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_BoxCoxAttributes.R
\name{ATA.BoxCoxAttr}
\alias{ATA.BoxCoxAttr}
\title{The ATA.BoxCoxAttr function works with many different types of inputs.}
\usage{
ATA.BoxCoxAttr(
  bcMethod = "loglik",
  bcLower = 0,
  bcUpper = 1,
  bcBiasAdj = FALSE
)
}
\arguments{
\item{bcMethod}{Choose method to be used in calculating lambda. "loglik" is default. Other method is "guerrero" (Guerrero, V.M. (1993)).}

\item{bcLower}{Lower limit for possible lambda values. The lower value is limited by -5. Default value is 0.}

\item{bcUpper}{Upper limit for possible lambda values. The upper value is limited by 5. Default value is 1.}

\item{bcBiasAdj}{Use adjusted back-transformed mean for Box-Cox transformations.
If transformed data is used to produce forecasts and fitted values, a regular back transformation will result in median forecasts.
If bcBiasAdj is TRUE, an adjustment will be made to produce mean forecasts and fitted values.
If bcBiasAdj=TRUE. Can either be the forecast variance, or a list containing the interval \code{level}, the corresponding \code{upper} and \code{lower} intervals.}
}
\value{
An object of class \code{ataattrset}.
}
\description{
The ATA.BoxCoxAttr function works with many different types of inputs.
}
\references{
Box, G. E. P. and Cox, D. R. (1964) An analysis of transformations. \emph{JRSS B} \bold{26} 211--246.

Guerrero, V.M. (1993) Time-series analysis supported by power transformations. \emph{Journal of Forecasting}, \bold{12}, 37--48.
}
\seealso{
\code{\link{BoxCox}}, \code{\link{InvBoxCox}}, \code{\link{BoxCox.lambda}}
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
