% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Forecast.R
\name{ATA.Forecast}
\alias{ATA.Forecast}
\title{Forecasting Time Series Using the ATA Method}
\usage{
ATA.Forecast(
  object,
  h = NULL,
  out.sample = NULL,
  ci.level = 95,
  negative.forecast = TRUE
)
}
\arguments{
\item{object}{An \code{ATA} object is required for forecast.}

\item{h}{Number of periods for forecasting.}

\item{out.sample}{A numeric vector or time series of class \code{ts} or \code{msts} for out-sample.}

\item{ci.level}{Confidence Interval levels for forecasting. Default value is 95.}

\item{negative.forecast}{Negative values are allowed for forecasting. Default value is TRUE. If FALSE, all negative values for forecasting are set to 0.}
}
\value{
An object of class "\code{ATA}".
}
\description{
\code{ATA.Forecast} is a generic function for forecasting of the ATA Method.
}
\examples{
ata.fit <- ATA(M3[[1899]]$x, M3[[1899]]$xx)
fc <- ATA.Forecast(ata.fit, h=18)
}
\references{
Yapar, G., (2016)
"Modified simple exponential smoothing"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi:10.15672/HJMS.201614320580

Yapar, G., Capar, S., Selamlar, H. T., Yavuz, I., (2016)
"Modified holt's linear trend method"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi: 10.15672/HJMS.2017.493
}
\seealso{
\code{forecast}, \code{stlplus}, \code{stR}, \code{\link[stats]{stl}}, \code{\link[stats]{decompose}},
\code{tbats}, \code{seasadj}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
\keyword{accuracy}
\keyword{ata}
\keyword{forecast}
\keyword{msts}
\keyword{ts}
