% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\name{ATA.plot}
\alias{ATA.plot}
\title{Specialized Plot Function of The ATAforecasting}
\usage{
ATA.plot(object, fcol = 4, flty = 2, flwd = 2, ...)
}
\arguments{
\item{object}{an object of \code{ATA}}

\item{fcol}{line color}

\item{flty}{line type}

\item{flwd}{line width}

\item{...}{other inputs}
}
\value{
a graphic output for the components of the ATAforecasting
}
\description{
Specialized Plot Function of The ATAforecasting
}
