% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\name{ATA.print}
\alias{ATA.print}
\title{Specialized Screen Print Function of The ATAforecasting}
\usage{
ATA.print(object, ...)
}
\arguments{
\item{object}{an object of \code{ATA}}

\item{...}{other inputs}
}
\value{
a summary for the results of the ATAforecasting
}
\description{
Specialized Screen Print Function of The ATAforecasting
}
