% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Main.r
\name{ATA}
\alias{ATA}
\alias{print.ata}
\alias{plot.ata}
\title{Forecasting Time Series by ATA Method with Box-Cox transformation}
\usage{
ATA(X, Y = NULL, parP = NULL, parQ = NULL, parPHI = NULL,
  model.type = NULL, seasonal.test = NULL, seasonal.model = NULL,
  seasonal.period = NULL, seasonal.type = NULL, seasonal.test.attr = NULL,
  find.period = NULL, accuracy.type = NULL, level.fixed = FALSE,
  trend.fixed = FALSE, h = NULL, partition.h = NULL,
  transform.method = NULL, lambda = NULL, initial.value = NULL,
  ci.level = 95, start.phi = NULL, end.phi = NULL, size.phi = NULL,
  negative.forecast = TRUE, plot.out = TRUE)
}
\arguments{
\item{X}{a numeric vector or time series of class \code{ts} or \code{msts} for in-sample.}

\item{Y}{a numeric vector or time series of class \code{ts} or \code{msts} for out-sample. If you do not have out-sample data, you can split in-sample data into training and test dataset with \code{partition.h} argument.}

\item{parP}{Value of Level parameter \code{p}. If NULL or "opt", it is estimated. \code{p} has all integer values from 1 to \code{length(X)}.}

\item{parQ}{Value of Trend parameter \code{q}. If NULL or "opt", it is estimated. \code{q} has all integer values from 0 to \code{p}.}

\item{parPHI}{Value of Damping Trend parameter \code{phi}. If NULL or "opt", it is estimated. phi has all values from 0 to 1.}

\item{model.type}{A one-character string identifying method using the framework terminology. The letter "A" for additive model, the letter "M" for multiplicative model. 
If NULL, both letters will be tried and the best model (according to the accuracy measure \code{accuracy.type}) returned.}

\item{seasonal.test}{Testing for stationary and seasonality. If TRUE, the method firstly uses \code{test="adf"}, Augmented Dickey-Fuller, unit-root test then the test returns the least number of differences required to pass the test at level \code{alpha}.
After the unit-root test, seasonal test applies on the stationary \code{X}.}

\item{seasonal.model}{A string identifying method for seasonal decomposition. If NULL, "decomp" method is default. c("none", "decomp", "stl", "stlplus", "tbats", "stR") phrases of methods denote
\itemize{
     \item{none}    : seasonal decomposition is not required.
     \item{decomp}     : classical seasonal decomposition. If \code{decomp}, the \code{stats} package will be used. 
     \item{stl}        : seasonal-trend decomposition procedure based on loess developed by Cleveland et al. (1990). If \code{stl}, the \code{stats} package will be used. 
     \item{stlplus}    : seasonal-trend decomposition procedure based on loess developed by Cleveland et al. (1990). If \code{stlplus}, the \code{stlplus} package will be used. 
     \item{tbats}   : exponential smoothing state space model with box-cox transformation, ARMA errors, trend and seasonal components 
                            as described in De Livera, Hyndman & Snyder (2011). Parallel processing is used by default to speed up the computations. If \code{tbats}, the \code{forecast} package will be used. 
     \item{stR}        : seasonal-trend decomposition procedure based on regression developed by Dokumentov and Hyndman (2015). If \code{stR}, the \code{stR} package will be used. 
     \item{x13}        : seasonal-trend decomposition procedure based on X13ARIMA/SEATS. If \code{x13}, the \code{seasonal} package will be used. 
     \item{x11}        : seasonal-trend decomposition procedure based on X11. If \code{x11}, the \code{seasonal} package will be used. 
}}

\item{seasonal.period}{Value(s) of seasonal periodicity. If NULL, \code{frequency} of X is default  If \code{seasonal.period} is not integer, \code{X} must be \code{msts} time series object. c(s1,s2,s3,...) for multiple period. If \code{X} has multiple periodicity, "tbats" or "stR" seasonal model have to be selected.}

\item{seasonal.type}{A one-character string identifying method for the seasonal component framework. If NULL, "M" is default. The letter "A" for additive model, the letter "M" for multiplicative model.
If other seasonal decomposition method except \code{decomp} with "M", Box-Cox transformation with \code{lambda}=0 is selected.}

\item{seasonal.test.attr}{Attributes set for unit root, seasonality tests, X13ARIMA/SEATS and X11. If NULL, s.tcrit=1.645, uroot.test="adf", uroot.alpha=0.05, uroot.maxd=2, x13.estimate.maxiter=1500, x13.estimate.tol=1.0e-5, x11.estimate.maxiter=1500, x11.estimate.tol=1.0e-5. If you want to change, please use \code{ata.seasonal.attr} function and its output.
For example, you can use \code{seasonal.test.attr = ata.seasonal.attr(s.tcrit=1.96)} equation in \code{ATA} function.}

\item{find.period}{Find seasonal period(s) automatically. If NULL, 0 is default. When \code{find.period},
\itemize{
     \item{0} : none
     \item{1} : single period with find.freq
     \item{2} : single period with \code{forecast::findfrequency}
     \item{3} : multiple period with find.freq & stR
     \item{4} : multiple period with find.freq & tbats
}}

\item{accuracy.type}{Accuracy measure for selection of the best model. IF NULL, \code{sMAPE} is default. 
\itemize{
     \item{MAE}        : mean absolute error.
     \item{MSE}        : mean square error.
     \item{RMSE}    : root mean squared error.
     \item{MPE}        : mean percentage error.
     \item{MAPE}    : mean absolute percentage error.
     \item{sMAPE}    : symmetric mean absolute percentage error.
     \item{MASE}    : mean absolute scaled error.
     \item{MdAE}    : median absolute error.
     \item{MdSE}    : median square error.
     \item{MdPE}    : median percentage error.
     \item{MdAPE}    : median absolute percentage error.
     \item{sMdAPE}    : symmetric median absolute percentage error.
}}

\item{level.fixed}{"pStarQ"  --> First, fits ATA(p,0) where p = p* is optimized for q=0. Then, fits ATA(p*,q) where q is optimized for p = p*.}

\item{trend.fixed}{"pBullet" --> Fits ATA(p,1) where p = p* is optimized for q = 1.}

\item{h}{The number of steps to forecast ahead.}

\item{partition.h}{If \code{Y} is NULL, this parameter divides \code{X} into two parts: training set (in-sample) and test set (out-sample). \code{partition.h} is number of periods for forecasting and size of test set. 
When the parameter is NULL; if the frequency of \code{X} is 4 the parameter is set to 8; if the frequency of \code{X} is 12 the parameter is set to 18; the parameter is set to 6 for other cases.}

\item{transform.method}{Transformation method  --> BoxCox, sqrt, inverse, log, log10. 
When \code{BoxCox} or \code{log} or \code{log10}  is specified,
\code{model.type} and \code{seasonal.type} is set to "A".}

\item{lambda}{Box-Cox transformation parameter. If NULL, data transformed before model is estimated. 
When \code{lambda} is specified, \code{model.type} and \code{seasonal.type} is set to "A".}

\item{initial.value}{If NULL, FALSE is default. If FALSE, ATA Method calculates the pth observation in \code{X} for level and qth observation in \code{X(T)-X(T-1)} for trend. 
If TRUE, ATA Method calculates average of first p value in \code{X}for level and average of first q value in \code{X(T)-X(T-1)} for trend.}

\item{ci.level}{Confidence Interval levels for forecasting.}

\item{start.phi}{Lower boundary for searching \code{parPHI}.If NULL, 0 is default.}

\item{end.phi}{Upper boundary for searching \code{parPHI}. If NULL, 1 is is default.}

\item{size.phi}{Increment step for searching \code{parPHI}. If NULL, 0.05 is default.}

\item{negative.forecast}{Negative values are allowed for forecasting. Default value is TRUE. If FALSE, all negative values for forecasting are set to 0.}

\item{plot.out}{Default is TRUE. If FALSE, graphics of ATA Method are not shown.}

\item{print.out}{Default is TRUE. If FALSE, summary of ATA Method is not shown.}

\item{...}{Other undocumented arguments.}

\item{actual}{The original time series.}

\item{fitted}{Fitted values (one-step forecasts). The mean is of the fitted values is calculated over the ensemble.}

\item{level}{Estimated level values.}

\item{trend}{Estimated trend values.}

\item{residuals}{Original values minus fitted values.}

\item{coefp}{The weights attached to level observations.}

\item{coefq}{The weights attached to trend observations.}

\item{p}{Optimum level parameter.}

\item{q}{Optimum trend parameter.}

\item{phi}{Optimum damped trend parameter.}

\item{model.type}{Form of trend.}

\item{h}{The number of steps to forecast ahead.}

\item{forecast}{Point forecasts as a time series.}

\item{out.sample}{Test values as a time series.}

\item{method}{The name of the optimum forecasting method as a character string.}

\item{initial.value}{Selected initial values for the time series forecasting method.}

\item{level.fixed}{A choice of optional level fixed trended methods.}

\item{trend.fixed}{A choice of optional trend fixed trended methods.}

\item{transform.method}{Transformation method  --> BoxCox, sqrt, inverse, log, log10.}

\item{lambda}{Box-Cox transformation parameter.}

\item{accuracy.type}{Accuracy measure that is chosen for model selection.}

\item{accuracy}{In and out sample accuracy measures and its descriptives that are calculated for optimum model are given.}

\item{is.season}{Indicates whether it contains seasonal pattern.}

\item{seasonal.model}{The name of the selected decomposition method.}

\item{seasonal.type}{Form of seasonality.}

\item{seasonal.period}{The number of seasonality periods (which defaults to \code{frequency(X)}).}

\item{seasonal.index}{Weights of seasonality.}

\item{seasonal}{Estimated seasonal values.}

\item{seasonal.adjusted}{Deseasonalized time series values.}

\item{execution.time}{The real and CPU time (in seconds) spent by the system executing that task, including the time spent executing run-time or system services on its behalf.}

\item{calculation.time}{How much real time (in seconds) the currently running R process has already taken.

The generic accessor functions \code{ATA.Forecast} and \code{ata.accuracy}
extract useful features of the value returned by \code{ata} and associated
functions.}
}
\value{
Returns an object of class "\code{ata}", containing.
}
\description{
Returns ATA(p,q,phi) applied to \code{X}.
Based on the modified simple exponential smoothing as described in Yapar, G. (2016). 
ATA method is a new univariate time series forecasting method which provides innovative 
solutions to issues faced during the initialization and optimization stages of existing methods. 
ATA's forecasting performance is superior to existing methods both in terms of easy implementation 
and accurate forecasting. It can be applied to non-seasonal or deseasonalized time series, 
where the deseasonalization can be performed via any preferred decomposition method.
This methodology performed extremely well on the M3 and M4-competition data.
}
\examples{
\dontrun{
fit <- ATA(M3[[1899]]$x,M3[[1899]]$xx)
plot(ATA.Forecast(fit,h=36))
}

}
\references{
Yapar, G., (2016)
"Modified simple exponential smoothing"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi:10.15672/HJMS.201614320580

Yapar, G., Capar, S., Selamlar, H. T., Yavuz, I., (2016) 
"Modified holt's linear trend method"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi: 10.15672/HJMS.2017.493
}
\seealso{
\code{\link{forecast}}, \code{\link{stlplus}}, \code{stR}, \code{\link[stats]{stl}}, \code{\link[stats]{decompose}},
\code{\link{tbats}}, \code{\link{seasadj}}, \code{\link{seasonal}}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
\keyword{accuracy}
\keyword{ata}
\keyword{forecast}
\keyword{msts}
\keyword{ts}
