% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Seasonality.r
\name{ATA.Seasonality}
\alias{ATA.Seasonality}
\title{Seasonality Tests for ATA Method}
\usage{
ATA.Seasonality(input, ppy, attr_set)
}
\arguments{
\item{input}{The data.}

\item{ppy}{Frequency of the data.}

\item{attr_set}{Assign from \code{ATA.SeasAttr} function. Attributes set for unit root, seasonality tests.}
}
\value{
\code{TRUE} if the serie has seasonality. Otherwise, \code{FALSE} .
}
\description{
Seasonality Tests for ATA Method
}
\references{
Dickey DA and Fuller WA (1979), "Distribution of the Estimators for
Autoregressive Time Series with a Unit Root", \code{Journal of the American
Statistical Association} \bold{74}:427-431.

Kwiatkowski D, Phillips PCB, Schmidt P and Shin Y (1992) "Testing the Null
Hypothesis of Stationarity against the Alternative of a Unit Root",
\code{Journal of Econometrics} \bold{54}:159-178.

Osborn DR, Chui APL, Smith J, and Birchenhall CR (1988) "Seasonality and the
order of integration for consumption", \code{Oxford Bulletin of Economics
and Statistics} \bold{50}(4):361-377.

Osborn, D.R. (1990) "A survey of seasonality in UK macroeconomic variables",
\code{International Journal of Forecasting}, \bold{6}:327-336.

Said E and Dickey DA (1984), "Testing for Unit Roots in Autoregressive
Moving Average Models of Unknown Order", \code{Biometrika}
\bold{71}:599-607.
Wang, X, Smith, KA, Hyndman, RJ (2006) "Characteristic-based clustering
for time series data", \code{Data Mining and Knowledge Discovery},
\bold{13}(3), 335-364.

Canova F and Hansen BE (1995) "Are Seasonal Patterns Constant
over Time? A Test for Seasonal Stability", \code{Journal of Business and
Economic Statistics} \bold{13}(3):237-252.

Hylleberg S, Engle R, Granger C and Yoo B (1990) "Seasonal integration
and cointegration.", \code{Journal of Econometrics} \bold{44}(1), pp. 215-238.
}
\seealso{
\code{\link{forecast}}, \code{\link{ucra}}, \code{\link{tseries}}, \code{\link{uroot}}, \code{\link{stlplus}}, \code{stR}, 
\code{\link[stats]{stl}}, \code{\link[stats]{decompose}}, \code{\link{tbats}}, \code{\link{seasadj}}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
