% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Shift.r
\name{ATA.Shift}
\alias{ATA.Shift}
\title{Lag/Lead (Shift) Function for Univariate Series}
\usage{
ATA.Shift(x, shift_by)
}
\arguments{
\item{x}{given variable(s)}

\item{shift_by}{lag or lead parameter}
}
\value{
Generating a lag/lead variables
}
\description{
Lag/Lead (Shift) Function for Univariate Series
}
