% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Shift.r
\name{ATA.Shift_Mat}
\alias{ATA.Shift_Mat}
\title{Lag/Lead (Shift) Function for Multivariate Series}
\usage{
ATA.Shift_Mat(X, direction = "down", shift_by = 1, fill = NA)
}
\arguments{
\item{X}{given matrice}

\item{direction}{direction of shifting. Default is "down".}

\item{shift_by}{number of rows to be shifed upwards/downwards}

\item{fill}{a value to be used to fill the rows}
}
\value{
Generating a lag/lead matrice
}
\description{
Lag/Lead (Shift) Function for Multivariate Series
}
