% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Transform.R
\name{ATA.Transform}
\alias{ATA.Transform}
\title{Transformation Techniques for ATA Method}
\usage{
ATA.Transform(X, tMethod = c("BoxCox", "BoxCox Shift", "Sqrt",
  "Sqrt Shift", "Reciprocal", "Log", "Log Shift", "NegLog", "Modulus",
  "Bickel-Doksum", "Manly", "Dual", "Yeo-Johnson", "GPower", "GLog"),
  tLambda, tShift = 0, bcMethod = c("loglik", "guerrero"),
  bcLower = 0, bcUpper = 1)
}
\arguments{
\item{X}{a numeric vector or time series of class \code{ts} or \code{msts} for in-sample.}

\item{tMethod}{Box-Cox power transformation family is consist of "BoxCox", "BoxCox Shift", "Sqrt", "Sqrt Shift", "Reciprocal", "Log", "Log Shift", "NegLog", "Modulus", "Bickel-Doksum", "Manly", "Dual", "Yeo-Johnson", "GPower", "GLog" in ATAforecasting package.}

\item{tLambda}{Box-Cox power transformation family parameter. If NULL, data transformed before model is estimated.}

\item{tShift}{Box-Cox power transformation family shifting parameter. If NULL, data transformed before model is estimated.}

\item{bcMethod}{Choose method to be used in calculating lambda. "loglik" is default. Other method is "guerrero" (Guerrero, V.M. (1993)).}

\item{bcLower}{Lower limit for possible lambda values. The lower value is limited by -5. Default value is 0.}

\item{bcUpper}{Upper limit for possible lambda values. The upper value is limited by 5. Default value is 1.}
}
\value{
A list object consists of transformation parameters and transformed data.
\code{ATA.Transform} is a list containing at least the following elements:
\itemize{
	 \item{trfmX}   : Transformed data
	 \item{tLambda} : Box-Cox power transformation family parameter
	 \item{tShift}  : Box-Cox power transformation family shifting parameter
}
}
\description{
These functions are modified version of \code{trafo::trafos} written by Lily Medina, Piedad Castro, Ann-Kristin Kreutzmann, Natalia Rojas-Perilla \code{trafo} package.
Please review manual and vignette documents of latest \code{trafo} package. The \code{ATA.Transform} function works with many different types of inputs.
}
