% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAforecasting-package.r
\docType{package}
\name{ATAforecasting-package}
\alias{ATAforecasting-package}
\alias{ATAforecasting}
\title{Forecasting Time Series by ATA Method with Box-Cox transformation}
\description{
Returns ATA(p,q,phi) applied to \code{X}.
Based on the modified simple exponential smoothing as described in Yapar, G. (2016). 
ATA method is a new univariate time series forecasting method which provides innovative 
solutions to issues faced during the initialization and optimization stages of existing methods. 
ATA's forecasting performance is superior to existing methods both in terms of easy implementation 
and accurate forecasting. It can be applied to non-seasonal or deseasonalized time series, 
where the deseasonalization can be performed via any preferred decomposition method.
This methodology performed extremely well on the M3 and M4-competition data.
}
\examples{
\dontrun{
fit <- ATA(M3[[1899]]$x,M3[[1899]]$xx)
plot(ATA.Forecast(fit,h=36))
}

}
\references{
Yapar, G., (2016)
"Modified simple exponential smoothing"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi:10.15672/HJMS.201614320580

Yapar, G., Capar, S., Selamlar, H. T., Yavuz, I., (2016) 
"Modified holt's linear trend method"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi: 10.15672/HJMS.2017.493
}
\seealso{
\code{\link{forecast}}, \code{\link{stlplus}}, \code{stR}, \code{\link[stats]{stl}}, \code{\link[stats]{decompose}},
\code{\link{tbats}}, \code{\link{seasadj}}, \code{\link{seasonal}}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
\keyword{accuracy}
\keyword{ata}
\keyword{forecast}
\keyword{msts}
\keyword{ts}
