% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Find_Freq_Fourier.R
\name{find.freq.fourier}
\alias{find.freq.fourier}
\title{Find Frequency Using Periodogram}
\usage{
find.freq.fourier(x)
}
\arguments{
\item{x}{an univariate time serie}
}
\value{
frequency/cycle of the given time serie
}
\description{
Find Frequency Using Periodogram
}
