% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Find_Freq.R
\name{find.freq}
\alias{find.freq}
\title{Find Frequency Using Spectral Density Of A Time Series From AR Fit}
\usage{
find.freq(x)
}
\arguments{
\item{x}{an univariate time series}
}
\value{
frequency/cycle of the given time data
}
\description{
Find Frequency Using Spectral Density Of A Time Series From AR Fit
}
