% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Find_Multi_Freq.R
\name{find.multi.freq}
\alias{find.multi.freq}
\title{Find Multi Frequency Using Spectral Density Of A Time Series From AR Fit}
\usage{
find.multi.freq(x)
}
\arguments{
\item{x}{an univariate time series}
}
\value{
multi frequencies/cycles of the given data
}
\description{
Find Multi Frequency Using Spectral Density Of A Time Series From AR Fit
}
