% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fundingTR}
\alias{fundingTR}
\title{Weekly Net Funding Level of Central Bank of Republic of Turkey}
\format{
Time series data
}
\source{
The Central Bank of the Republic of Turkey -- CBRT.
}
\usage{
data(fundingTR)
}
\description{
Weekly Net Funding Level of Central Bank of Republic of Turkey:  from Jan 7, 2011 to Jan 08, 2021.
}
\examples{
plot(fundingTR)

}
\keyword{datasets}
