% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Plot.R
\name{plot.ata}
\alias{plot.ata}
\title{Specialized Plot Function of The ATA Method}
\usage{
plot.ata(x, fcol = 4, flty = 2, flwd = 2, ...)
}
\arguments{
\item{x}{an object of \code{ata}}

\item{fcol}{line color}

\item{flty}{line type}

\item{flwd}{line width}

\item{...}{}
}
\value{
a graphic output for the components of the ATA Methods
}
\description{
Specialized Plot Function of The ATA Method
}
