% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Print.R
\name{print.ata}
\alias{print.ata}
\title{Specialized Screen Print Function of The ATA Method}
\usage{
print.ata(x, ...)
}
\arguments{
\item{x}{an object of \code{ata}}

\item{...}{other inputs}
}
\value{
a summary for the results of the ATA Methods
}
\description{
Specialized Screen Print Function of The ATA Method
}
