% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Decomposition.R
\name{stR_seasadj}
\alias{stR_seasadj}
\title{Seasonal adjustment based on STR}
\usage{
stR_seasadj(object, include = c("Trend", "Random"))
}
\arguments{
\item{object}{Result of \code{AutoSTR} function in ATA.Decomposition}

\item{include}{Vector of component names to include in the result. The default is \code{c("Trend", "Random")}.}
}
\value{
Seasonal adjusted data using stR package.
}
\description{
This function is written to avoid conflict two functions with the same name, \code{seasadj}, in two different packages: \code{forecast} and \code{stR}.
}
\keyword{decomposition}
\keyword{seasonal}
\keyword{stR}
