% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{touristTR}
\alias{touristTR}
\title{Monthly number of tourists arrived in Turkey}
\format{
Time series data
}
\source{
The Central Bank of the Republic of Turkey -- CBRT.
}
\usage{
data(touristTR)
}
\description{
Monthly number of tourists arrived in Turkey: from Jan 2008 to Dec 2020.
}
\examples{
plot(touristTR)

}
\keyword{datasets}
