% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_CI.r
\name{ATA.CI}
\alias{ATA.CI}
\title{Confidence Interval function for the ATA Method}
\usage{
ATA.CI(object, ci.level = 95)
}
\arguments{
\item{object}{An \code{ATA} object is required.}

\item{ci.level}{Confidence level, for example: 90, 95 or 99.}
}
\value{
The confidence interval output for the ATA forecasts
}
\description{
Confidence Interval function for the ATA Method
}
