% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Forecast.r
\name{ATA.Forecast}
\alias{ATA.Forecast}
\title{Forecasting time series using ATA Method}
\usage{
ATA.Forecast(y, h = NULL, out.sample = NULL, ci.alpha = 0.95)
}
\arguments{
\item{y}{An \code{ata} object is required for forecast.}

\item{h}{Number of periods for forecasting.}

\item{out.sample}{A numeric vector or time series of class \code{ts} or \code{msts} for out-sample.}

\item{ci.level}{Confidence Interval levels for forecasting.}
}
\value{
An object of class "\code{forecast}".
}
\description{
\code{ATA.Forecast} is a generic function for forecasting of ATA Method. 
The function invokes particular \emph{methods} which
depend on the class of the first argument.
}
\examples{

ata.fit <- ATA(window(WWWusage, end=60))
fc <- ATA.Forecast(WWWusage, model=fit)

}
\references{
Yapar, G., (2016)
"Modified simple exponential smoothing"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi:10.15672/HJMS.201614320580

Yapar, G., Capar, S., Selamlar, H. T., Yavuz, I., (2016) 
"Modified holt's linear trend method"
\emph{Hacettepe University Journal of Mathematics and Statistics} Early Access. Doi: 10.15672/HJMS.2017.493
}
\seealso{
\code{\link{forecast}}, \code{\link{stlplus}}, \code{stR}, \code{\link[stats]{stl}}, \code{\link[stats]{decompose}},
\code{\link{tbats}}, \code{\link{seasadj}}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
\keyword{accuracy}
\keyword{ata}
\keyword{forecast}
\keyword{msts}
\keyword{ts}
