% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\name{ATA.Print}
\alias{ATA.Print}
\title{Specialized Screen Print Function of The ATAforecasting}
\usage{
ATA.Print(object, ...)
}
\arguments{
\item{object}{an object of \code{ata}}

\item{...}{other inputs}
}
\value{
a summary for the results of the ATAforecasting
}
\description{
Specialized Screen Print Function of The ATAforecasting
}
