% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_Seasonality.R
\name{ATA.Seasonality}
\alias{ATA.Seasonality}
\title{Seasonality Tests for The ATA Method}
\usage{
ATA.Seasonality(input, ppy, attr_set)
}
\arguments{
\item{input}{The data.}

\item{ppy}{Frequency of the data.}

\item{attr_set}{Assign from \code{ATA.SeasAttr} function. Attributes set for unit root, seasonality tests.}
}
\value{
\code{TRUE} if the serie has seasonality. Otherwise, \code{FALSE}.
}
\description{
This function is a class of seasonality tests using  \code{corrgram.test} from ATAforecasting package, \code{ndiffs} and \code{nsdiffs} functions from forecast package.
Also, this function is modified version of \code{ndiffs} and \code{nsdiffs} written by Hyndman et al. \code{forecast} package.
Please review manual and vignette documents of latest \code{forecast} package. According to \code{forecast} package,
\code{ndiffs} and \code{nsdiffs} functions to estimate the number of differences required to make a given time series stationary.
\code{ndiffs} uses unit root tests to determine the number of differences required for time series to be made trend stationary. Several different tests are available:
\itemize{
	\item {uroot.test = 'kpss'}			: the KPSS test is used with the null hypothesis that \code{x} has a stationary root against a unit-root alternative. Then the test returns the least number of differences required to pass the test at the level \code{uroot.alpha}.
	\item {uroot.test = 'adf'}			: the Augmented Dickey-Fuller test is used.
	\item {uroot.test = 'pp'}			: the Phillips-Perron test is used. In both of these cases, the null hypothesis is that \code{x} has a unit root against a stationary root alternative. Then the test returns the least number of differences required to fail the test at the level \code{alpha}.
}
\code{nsdiffs} uses seasonal unit root tests to determine the number of seasonal differences required for time series to be made stationary. Several different tests are available:
\itemize{
	\item {suroot.test = 'seas'}		: a measure of seasonal strength is used, where differencing is selected if the seasonal strength (Wang, Smith & Hyndman, 2006) exceeds 0.64 (based on minimizing MASE when forecasting using auto.arima on M3 and M4 data).
	\item {suroot.test = 'ch'}			: the Canova-Hansen (1995) test is used (with null hypothesis of deterministic seasonality)
	\item {suroot.test = 'hegy'}		: the Hylleberg, Engle, Granger & Yoo (1990) test is used.
	\item {suroot.test = 'ocsb'}		: the Osborn-Chui-Smith-Birchenhall (1988) test is used (with null hypothesis that a seasonal unit root exists).
	\item {suroot.test = 'correlogram'}	: this function is written based on M4 Competition Seasonality Test.
}
}
\references{
Dickey DA and Fuller WA (1979), "Distribution of the Estimators for
Autoregressive Time Series with a Unit Root", \code{Journal of the American
Statistical Association} \bold{74}:427-431.

Kwiatkowski D, Phillips PCB, Schmidt P and Shin Y (1992) "Testing the Null
Hypothesis of Stationarity against the Alternative of a Unit Root",
\code{Journal of Econometrics} \bold{54}:159-178.

Osborn DR, Chui APL, Smith J, and Birchenhall CR (1988) "Seasonality and the
order of integration for consumption", \code{Oxford Bulletin of Economics
and Statistics} \bold{50}(4):361-377.

Osborn, D.R. (1990) "A survey of seasonality in UK macroeconomic variables",
\code{International Journal of Forecasting}, \bold{6}:327-336.

Said E and Dickey DA (1984), "Testing for Unit Roots in Autoregressive
Moving Average Models of Unknown Order", \code{Biometrika}
\bold{71}:599-607.
Wang, X, Smith, KA, Hyndman, RJ (2006) "Characteristic-based clustering
for time series data", \code{Data Mining and Knowledge Discovery},
\bold{13}(3), 335-364.

Canova F and Hansen BE (1995) "Are Seasonal Patterns Constant
over Time? A Test for Seasonal Stability", \code{Journal of Business and
Economic Statistics} \bold{13}(3):237-252.

Hylleberg S, Engle R, Granger C and Yoo B (1990) "Seasonal integration
and cointegration.", \code{Journal of Econometrics} \bold{44}(1), pp. 215-238.
}
\seealso{
\code{\link{forecast}}, \code{\link{ucra}}, \code{\link{tseries}}, \code{\link{uroot}}, \code{\link{stlplus}}, \code{stR},
\code{\link[stats]{stl}}, \code{\link[stats]{decompose}}, \code{\link{tbats}}, \code{\link{seasadj}}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar
}
