% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\docType{package}
\name{ATAforecasting-package}
\alias{ATAforecasting-package}
\title{ATAforecasting: Forecasting Time Series by ATA Method with Box-Cox Power Transformations Family and Seasonal Decomposition Techniques}
\description{
Returns ATA(p,q,phi) applied to \code{X}.
Based on the modified simple exponential smoothing as described in Yapar, G. (2016).
ATA method is a new univariate time series forecasting method which provides innovative
solutions to issues faced during the initialization and optimization stages of existing methods.
ATA's forecasting performance is superior to existing methods both in terms of easy implementation
and accurate forecasting. It can be applied to non-seasonal or deseasonalized time series,
where the deseasonalization can be performed via any preferred decomposition method.
This methodology performed extremely well on the M3 and M4-competition data.
Returns ATA(p,q,phi) applied to \code{X}.
}
\author{
Ali Sabri Taylan and Hanife Taylan Selamlar

Maintainer: alisabritaylan@gmail.com
}
\keyword{package}
