% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA.R
\name{print.ATA}
\alias{print.ATA}
\title{Specialized Screen Print Function of The ATA Method Forecast}
\usage{
print.ATA(object, ...)
}
\arguments{
\item{object}{an object of \code{ATA}}

\item{...}{other inputs}
}
\value{
a summary for the results of the ATA Methods
}
\description{
Specialized Screen Print Function of The ATA Method Forecast
}
